Add settable max activity amount

+++ src/GRA.Domain.Repository/IUserLogRepository.cs
@@ -19,7 +19,7 @@ namespace GRA.Domain.Repository
             ICollection<int?> translationIds);
         Task<ICollection<int>> UserIdsEarnedBadgeAsync(int badgeId, ReportCriterion criterion);
         Task<ICollection<int>> UserIdsCompletedChallengesAsync(int challengeId, ReportCriterion criterion);
-        Task<long> GetActivityEarnedForUserAsync(int userId);
+        Task<int> GetActivityEarnedForUserAsync(int userId);
         Task<bool> PointTranslationHasBeenUsedAsync(int translationId);
     }
 }
